Каждый класс реализован в отдельном файле. Интернет-магазин работает после сборки с Webpack
