+++ public/scripts/shoppack.js
@@ -1,1 +1,1 @@
-(()=>{"use strict";function t(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class e{constructor(e,s,r){t(this,"_id",void 0),t(this,"_name",void 0),t(this,"_price",void 0),this._id=e,this._name=s,this._price=r}get ID(){return this._id}get Name(){return this._name}get Price(){return this._price}}class s{constructor(){var t,e;e=[],(t="Items")in this?Object.defineProperty(this,t,{value:e,enumerable:!0,configurable:!0,writable:!0}):this[t]=e}getIndexFromID(t){if(0==this.Items.length)return-1;let e=0;for(;e<this.Items.length;){if(this.Items[e].ID==t)return e;e++}return-1}}function r(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class i extends s{constructor(t,e){super(),r(this,"_chunkSize",void 0),r(this,"_currentItems",void 0),r(this,"_querySelectorName",void 0),this._querySelectorName=e,this._chunkSize=t,this._currentItems=t}get ChunkSize(){return this._chunkSize}get CurrentItems(){return this._currentItems}get QuerySelectorName(){return this._querySelectorName}_getProducts(){let t=`${document.location.href}database/products.json`;return fetch(t).then((t=>t.json())).then((t=>t.products)).catch((()=>{reject(new Error("Ошибка при получении массива свойств товаров!"))}))}getProductsList(){return new Promise(((t,s)=>{this._getProducts().then((s=>{s=s.map((t=>new e(t.id,t.name,t.price))),this.Items.push(...s),t()})).catch((()=>{s(new Error("Ошибка при обработке массива свойств товаров!"))}))}))}incCurrentItems(){this._currentItems+this._chunkSize<=this.Items.length?this._currentItems+=this._chunkSize:this._currentItems=this.Items.length}_createProductsList(){let t="";for(let e=0;e<this.CurrentItems;e++)t+=`<p>${this.Items[e].Name}, цена: ${this.Items[e].Price} <button type="button" name="btnProductList_${this.Items[e].ID}" value="${this.Items[e].ID}">+</button></p><br>`;return t}render(){let t=document.querySelector("."+this._querySelectorName);if(null!=t){const e=document.createElement("text");e.classList.add("Products");let s=this._createProductsList();t.innerHTML="",e.innerHTML=s,t.appendChild(e)}}}class n extends e{constructor(t,e,s){var r,i;super(t,e,s),i=void 0,(r="_count")in this?Object.defineProperty(this,r,{value:i,enumerable:!0,configurable:!0,writable:!0}):this[r]=i,this._count=0}get Count(){return this._count}incCount(){this._count++}decCount(){this._count>0&&this._count--}}function o(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class c extends s{constructor(t,e){super(),o(this,"_inetShopProductsList",void 0),o(this,"_querySelectorName",void 0),this._inetShopProductsList=t,this._querySelectorName=e}get InetShopProductsList(){return this._inetShopProductsList}addProduct(t){let e=this.getIndexFromID(t);if(-1!=e)return this.Items[e].incCount(),e;e=this._inetShopProductsList.getIndexFromID(t);const s=new n(this._inetShopProductsList.Items[e].ID,this._inetShopProductsList.Items[e].Name,this._inetShopProductsList.Items[e].Price);return s.incCount(),this.Items.push(s),this.Items.length-1}deleteProductElement(t){const e=this.getIndexFromID(t);-1!=e&&(this.Items[e].decCount(),0==this.Items[e].Count&&this.Items.splice(e,1))}deleteProduct(t){const e=this.getIndexFromID(t);-1!=e&&this.Items.splice(e,1)}deleteAllProducts(){this.Items.splice(0,this.Items.length)}_createBasketProductsList(){let t="";return this.Items.forEach((e=>{let s=`<p>${e.Name}, цена: ${e.Price}, кол-во: ${e.Count} <button type = "button" name="btnDeleteElement_${e.ID}" value="${e.ID}">-</button> <button type = "button" name="btnDeleteProduct_${e.ID}" value="${e.ID}">:-(</button></p><br>`;t+=s})),t}render(){let t=document.querySelector("."+this._querySelectorName);if(null!=t){const e=document.createElement("text");e.classList.add("BasketProducts"),t.innerHTML="";let s=this._createBasketProductsList();e.innerHTML=s,t.appendChild(e)}}}function u(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class d{constructor(t,e){u(this,"_countTotal",0),u(this,"_sumTotal",0),u(this,"_querySelectorName",void 0),this._basketProductsList=t,this._querySelectorName=e}get BasketProductsList(){return this._basketProductsList}get CountTotal(){return this._countTotal}get SumTotal(){return this._sumTotal}_calculate(){this._countTotal=0,this._sumTotal=0,0!=this._basketProductsList.Items.length&&this._basketProductsList.Items.forEach((t=>{this._countTotal+=t.Count,this._sumTotal+=t.Price*t.Count}))}render(){this._calculate();let t=document.querySelector("."+this._querySelectorName);if(null!=t){const e=document.createElement("text");e.classList.add("BasketInfo"),t.innerHTML="";let s=`Количество товаров: ${this.CountTotal} на сумму: ${this.SumTotal}`;e.innerHTML=s,t.appendChild(e)}}}function h(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class l{constructor(t,e){h(this,"_basketProductsList",void 0),h(this,"_basket",void 0),this._basketProductsList=t,this._basket=e}get BasketProductsList(){return this._basketProductsList}get Basket(){return this._basket}click(){this._basketProductsList.render(),this._basket.render()}}class a extends l{constructor(t,e,s,r){super(s,r),h(this,"_productID",void 0),h(this,"_inetShopProductsList",void 0),h(this,"_index",void 0),this._productID=t,this._inetShopProductsList=e,this._index=this._inetShopProductsList.getIndexFromID(this._productID)}get ProductID(){return this._productID}get InetShopProductsList(){return this._interShopProductsList}get Index(){return this._index}click(){this.BasketProductsList.addProduct(this._productID),super.click()}}class m extends l{constructor(t,e){super(t,e)}click(){super.BasketProductsList.deleteAllProducts(),super.click()}}class _ extends l{constructor(t,e,s){super(e,s),h(this,"_id",void 0),this._id=t}get ID(){return this._id}click(){super.BasketProductsList.deleteProductElement(this._id),super.click()}}class P extends l{constructor(t,e,s){super(e,s),h(this,"_id",void 0),this._id=t}get ID(){return this._id}click(){super.BasketProductsList.deleteProduct(this._id),super.click()}}class b{static _prepareShopObjects(){this.interShopProductsList=new i(2,"productsList"),this.basketProductsList=new c(this.interShopProductsList,"basketList"),this.basket=new d(this.basketProductsList,"basket")}static _addtoBasket(t){return new Promise(((e,s)=>{new a(t,this.interShopProductsList,this.basketProductsList,this.basket).click(),e()}))}static OnBtnAddToBasketClick(t){this._addtoBasket(t).then((()=>{this._setHandlersForBasketButtons()}))}static _deleteProductElement(t){return new Promise(((e,s)=>{new _(t,this.basketProductsList,this.basket).click(),e()}))}static OnBtnDeleteProductElementClick(t){this._deleteProductElement(t).then((()=>{this._setHandlersForBasketButtons()}))}static _deleteProduct(t){return new Promise(((e,s)=>{new P(t,this.basketProductsList,this.basket).click(),e()}))}static OnBtnDeleteProductClick(t){this._deleteProduct(t).then((()=>{this._setHandlersForBasketButtons()}))}static _deleteAllProducts(){return new Promise(((t,e)=>{new m(this.basketProductsList,this.basket).click(),t()}))}static OnBtnDeleteAllProducts(){this._deleteAllProducts().then((()=>{this.basket.render()}))}static OnBtnGetNextChunkProducts(){this.interShopProductsList.incCurrentItems(),this.interShopProductsList.CurrentItems<=this.interShopProductsList.Items.length&&(this.interShopProductsList.render(),this._setHandlerForProductsListButtons())}static _setHandlerForProductsListButtons(){for(let t=0;t<this.interShopProductsList.CurrentItems;t++){let e=this.interShopProductsList.Items[t].ID,s=document.getElementsByName(this._btnInetShopProductsName+"_"+e)[0];null!=s?s.addEventListener("click",(()=>{this.OnBtnAddToBasketClick(s.value)})):console.error("Нет конпки "+this._btnInetShopProductsName+"_"+e)}}static _setHandlersForBasketButtons(){this.basketProductsList.Items.forEach((t=>{let e=t.ID,s=document.getElementsByName(this._btnDeleteProductElementName+"_"+e)[0];null!=s&&s.addEventListener("click",(()=>{this.OnBtnDeleteProductElementClick(s.value)})),s=document.getElementsByName(this._btnDeleteProductName+"_"+e)[0],null!=s&&s.addEventListener("click",(()=>{this.OnBtnDeleteProductClick(s.value)}))}))}static _setHandlerForDeleteAllProductsButtons(){let t=document.getElementsByName(this._btnDeleteAllProductsName)[0];null!=t&&t.addEventListener("click",(()=>{this.OnBtnDeleteAllProducts()}))}static _setHandlerForGetNextChunkProductsButtons(){let t=document.getElementsByName(this._btnGetNextChunkName)[0];null!=t&&t.addEventListener("click",(()=>{this.OnBtnGetNextChunkProducts()}))}static main(){try{this._prepareShopObjects(),this.interShopProductsList.getProductsList().then((()=>{this.interShopProductsList.render(),this._setHandlerForProductsListButtons()})).catch((t=>{console.error("Основной модуль!",t)})),this._setHandlerForGetNextChunkProductsButtons(),this.basket.render(),this._setHandlerForDeleteAllProductsButtons()}catch(t){console.error(t)}}}h(b,"_btnInetShopProductsName","btnProductList"),h(b,"_btnDeleteProductElementName","btnDeleteElement"),h(b,"_btnDeleteProductName","btnDeleteProduct"),h(b,"_btnDeleteAllProductsName","ClearBasket"),h(b,"_btnGetNextChunkName","btnAddNextChunk"),h(b,"interShopProductsList",void 0),h(b,"basketProductsList",void 0),h(b,"basket",void 0),b.main()})();
+(()=>{"use strict";function t(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class e{constructor(e,s,r){t(this,"_id",void 0),t(this,"_name",void 0),t(this,"_price",void 0),this._id=e,this._name=s,this._price=r}get ID(){return this._id}get Name(){return this._name}get Price(){return this._price}}class s{constructor(){var t,e;e=[],(t="Items")in this?Object.defineProperty(this,t,{value:e,enumerable:!0,configurable:!0,writable:!0}):this[t]=e}getIndexFromID(t){if(0==this.Items.length)return-1;let e=0;for(;e<this.Items.length;){if(this.Items[e].ID==t)return e;e++}return-1}}function r(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class i extends s{constructor(t,e){super(),r(this,"_chunkSize",void 0),r(this,"_currentItems",void 0),r(this,"_querySelectorName",void 0),this._querySelectorName=e,this._chunkSize=t,this._currentItems=t}get ChunkSize(){return this._chunkSize}get CurrentItems(){return this._currentItems}get QuerySelectorName(){return this._querySelectorName}_getProducts(){let t=`${document.location.href}database/products.json`;return fetch(t).then((t=>t.json())).then((t=>t.products)).catch((()=>{reject(new Error("Ошибка при получении массива свойств товаров!"))}))}getProductsList(){return new Promise(((t,s)=>{this._getProducts().then((s=>{s=s.map((t=>new e(t.id,t.name,t.price))),this.Items.push(...s),t()})).catch((()=>{s(new Error("Ошибка при обработке массива свойств товаров!"))}))}))}incCurrentItems(){this._currentItems+this._chunkSize<=this.Items.length?this._currentItems+=this._chunkSize:this._currentItems=this.Items.length}_createProductsList(){let t="";for(let e=0;e<this.CurrentItems;e++)t+=`<p>${this.Items[e].Name}, цена: ${this.Items[e].Price} <button type="button" name="btnProductList_${this.Items[e].ID}" value="${this.Items[e].ID}">+</button></p><br>`;return t}render(){let t=document.querySelector("."+this._querySelectorName);if(null!=t){const e=document.createElement("text");e.classList.add("Products");let s=this._createProductsList();t.innerHTML="",e.innerHTML=s,t.appendChild(e)}}}class n extends e{constructor(t,e,s){var r,i;super(t,e,s),i=void 0,(r="_count")in this?Object.defineProperty(this,r,{value:i,enumerable:!0,configurable:!0,writable:!0}):this[r]=i,this._count=0}get Count(){return this._count}incCount(){this._count++}decCount(){this._count>0&&this._count--}}function o(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class c extends s{constructor(t,e){super(),o(this,"_inetShopProductsList",void 0),o(this,"_querySelectorName",void 0),this._inetShopProductsList=t,this._querySelectorName=e}get InetShopProductsList(){return this._inetShopProductsList}addProduct(t){let e=this.getIndexFromID(t);if(-1!=e)return this.Items[e].incCount(),e;e=this._inetShopProductsList.getIndexFromID(t);const s=new n(this._inetShopProductsList.Items[e].ID,this._inetShopProductsList.Items[e].Name,this._inetShopProductsList.Items[e].Price);return s.incCount(),this.Items.push(s),this.Items.length-1}deleteProductElement(t){const e=this.getIndexFromID(t);-1!=e&&(this.Items[e].decCount(),0==this.Items[e].Count&&this.Items.splice(e,1))}deleteProduct(t){const e=this.getIndexFromID(t);-1!=e&&this.Items.splice(e,1)}deleteAllProducts(){this.Items.splice(0,this.Items.length)}_createBasketProductsList(){let t="";return this.Items.forEach((e=>{let s=`<p>${e.Name}, цена: ${e.Price}, кол-во: ${e.Count} <button type = "button" name="btnDeleteElement_${e.ID}" value="${e.ID}">-</button> <button type = "button" name="btnDeleteProduct_${e.ID}" value="${e.ID}">:-(</button></p><br>`;t+=s})),t}render(){let t=document.querySelector("."+this._querySelectorName);if(null!=t){const e=document.createElement("text");e.classList.add("BasketProducts"),t.innerHTML="";let s=this._createBasketProductsList();e.innerHTML=s,t.appendChild(e)}}}function u(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class d{constructor(t,e){u(this,"_countTotal",0),u(this,"_sumTotal",0),u(this,"_querySelectorName",void 0),this._basketProductsList=t,this._querySelectorName=e}get BasketProductsList(){return this._basketProductsList}get CountTotal(){return this._countTotal}get SumTotal(){return this._sumTotal}_calculate(){this._countTotal=0,this._sumTotal=0,0!=this._basketProductsList.Items.length&&this._basketProductsList.Items.forEach((t=>{this._countTotal+=t.Count,this._sumTotal+=t.Price*t.Count}))}render(){this._calculate();let t=document.querySelector("."+this._querySelectorName);if(null!=t){const e=document.createElement("text");e.classList.add("BasketInfo"),t.innerHTML="";let s=`Количество товаров: ${this.CountTotal} на сумму: ${this.SumTotal}`;e.innerHTML=s,t.appendChild(e)}}}function l(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class h{constructor(t,e){l(this,"_basketProductsList",void 0),l(this,"_basket",void 0),this._basketProductsList=t,this._basket=e}get BasketProductsList(){return this._basketProductsList}get Basket(){return this._basket}click(){this._basketProductsList.render(),this._basket.render()}}function a(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class m extends h{constructor(t,e,s,r){super(s,r),a(this,"_productID",void 0),a(this,"_inetShopProductsList",void 0),a(this,"_index",void 0),this._productID=t,this._inetShopProductsList=e,this._index=this._inetShopProductsList.getIndexFromID(this._productID)}get ProductID(){return this._productID}get InetShopProductsList(){return this._interShopProductsList}get Index(){return this._index}click(){this.BasketProductsList.addProduct(this._productID),super.click()}}class _ extends h{constructor(t,e){super(t,e)}click(){super.BasketProductsList.deleteAllProducts(),super.click()}}class P extends h{constructor(t,e,s){var r;super(e,s),r=void 0,"_id"in this?Object.defineProperty(this,"_id",{value:r,enumerable:!0,configurable:!0,writable:!0}):this._id=r,this._id=t}get ID(){return this._id}click(){super.BasketProductsList.deleteProductElement(this._id),super.click()}}class b extends h{constructor(t,e,s){var r;super(e,s),r=void 0,"_id"in this?Object.defineProperty(this,"_id",{value:r,enumerable:!0,configurable:!0,writable:!0}):this._id=r,this._id=t}get ID(){return this._id}click(){super.BasketProductsList.deleteProduct(this._id),super.click()}}function p(t,e,s){return e in t?Object.defineProperty(t,e,{value:s,enumerable:!0,configurable:!0,writable:!0}):t[e]=s,t}class k{static _prepareShopObjects(){this.interShopProductsList=new i(2,"productsList"),this.basketProductsList=new c(this.interShopProductsList,"basketList"),this.basket=new d(this.basketProductsList,"basket")}static _addtoBasket(t){return new Promise(((e,s)=>{new m(t,this.interShopProductsList,this.basketProductsList,this.basket).click(),e()}))}static OnBtnAddToBasketClick(t){this._addtoBasket(t).then((()=>{this._setHandlersForBasketButtons()}))}static _deleteProductElement(t){return new Promise(((e,s)=>{new P(t,this.basketProductsList,this.basket).click(),e()}))}static OnBtnDeleteProductElementClick(t){this._deleteProductElement(t).then((()=>{this._setHandlersForBasketButtons()}))}static _deleteProduct(t){return new Promise(((e,s)=>{new b(t,this.basketProductsList,this.basket).click(),e()}))}static OnBtnDeleteProductClick(t){this._deleteProduct(t).then((()=>{this._setHandlersForBasketButtons()}))}static _setHandlersForBasketButtons(){this.basketProductsList.Items.forEach((t=>{let e=t.ID,s=document.getElementsByName(this._btnDeleteProductElementName+"_"+e)[0];null!=s&&s.addEventListener("click",(()=>{this.OnBtnDeleteProductElementClick(s.value)})),s=document.getElementsByName(this._btnDeleteProductName+"_"+e)[0],null!=s&&s.addEventListener("click",(()=>{this.OnBtnDeleteProductClick(s.value)}))}))}static _deleteAllProducts(){return new Promise(((t,e)=>{new _(this.basketProductsList,this.basket).click(),t()}))}static OnBtnDeleteAllProducts(){this._deleteAllProducts().then((()=>{this.basket.render()}))}static OnBtnGetNextChunkProducts(){this.interShopProductsList.incCurrentItems(),this.interShopProductsList.CurrentItems<=this.interShopProductsList.Items.length&&(this.interShopProductsList.render(),this._setHandlerForProductsListButtons())}static _setHandlerForProductsListButtons(){for(let t=0;t<this.interShopProductsList.CurrentItems;t++){let e=this.interShopProductsList.Items[t].ID,s=document.getElementsByName(this._btnInetShopProductsName+"_"+e)[0];null!=s?s.addEventListener("click",(()=>{this.OnBtnAddToBasketClick(s.value)})):console.error("Нет конпки "+this._btnInetShopProductsName+"_"+e)}}static _setHandlerForDeleteAllProductsButtons(){let t=document.getElementsByName(this._btnDeleteAllProductsName)[0];null!=t&&t.addEventListener("click",(()=>{this.OnBtnDeleteAllProducts()}))}static _setHandlerForGetNextChunkProductsButtons(){let t=document.getElementsByName(this._btnGetNextChunkName)[0];null!=t&&t.addEventListener("click",(()=>{this.OnBtnGetNextChunkProducts()}))}static main(){try{this._prepareShopObjects(),this.interShopProductsList.getProductsList().then((()=>{this.interShopProductsList.render(),this._setHandlerForProductsListButtons()})).catch((t=>{console.error("Основной модуль!",t)})),this._setHandlerForGetNextChunkProductsButtons(),this.basket.render(),this._setHandlerForDeleteAllProductsButtons()}catch(t){console.error(t)}}}p(k,"_btnInetShopProductsName","btnProductList"),p(k,"_btnDeleteProductElementName","btnDeleteElement"),p(k,"_btnDeleteProductName","btnDeleteProduct"),p(k,"_btnDeleteAllProductsName","ClearBasket"),p(k,"_btnGetNextChunkName","btnAddNextChunk"),p(k,"interShopProductsList",void 0),p(k,"basketProductsList",void 0),p(k,"basket",void 0),k.main()})();
